% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DPdata_missing}
\alias{DPdata_missing}
\title{Generated dataset of Dot-probe data with missing data}
\format{A dataframe with 3840 rows and 6 variables
\itemize{
\item subject: contains participant numbers for 20 subjects
\item blockcode: two block conditions "block1" and "block2"
\item trialnum: 96 trials per block
\item congruency: sets to congruent or incongruent trials
\item latency: RT measure (simulated data)
\item correct: accuracy (set to all accurate for the example)
}}
\usage{
DPdata_missing
}
\description{
A dataset containing data necessary to run examples of each function
including missing data
}
\keyword{datasets}
