% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohter.R
\name{other}
\alias{other}
\title{Title}
\usage{
other(df.split, method, sub, var1, var2, var3)
}
\arguments{
\item{df.split}{被分半的list}

\item{method}{分半方法}

\item{sub}{被试}

\item{var1}{变量1}

\item{var2}{变量2}

\item{var3}{变量3}
}
\value{
返回分半的结果
}
\description{
Title
}
