% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_tool.R
\name{sh_tool}
\alias{sh_tool}
\title{Title}
\usage{
sh_tool(df.split, iteration, nc, sub, var1, var2, var3, method)
}
\arguments{
\item{df.split}{data}

\item{iteration}{iteration}

\item{nc}{nunmber of cores (only for mc)}

\item{sub}{Subject}

\item{var1}{variable 1}

\item{var2}{variable 2}

\item{var3}{variable 3}

\item{method}{split method, "mc", "fs", "od", "permuted"}
}
\value{
返回分半的结果
}
\description{
Title
}
