% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSTdata_missing}
\alias{TSTdata_missing}
\title{Generated dataset of Task switching data with missing data}
\format{A dataframe with 3840 rows and 6 variables
\itemize{
\item subject: contains participant numbers for 20 subjects
\item blockcode: two block conditions "block1" and "block2"
\item trialnum: 96 trials per block
\item trialtype: sets to repeat or switch trials
\item latency: RT measure (simulated data)
\item correct: accuracy (set to all accurate for the example)
}}
\usage{
TSTdata_missing
}
\description{
The data is adapted from the DPdata_missing set using the following code
}
\details{
TSTdata_missing <- DPdata_missing
names(TSTdata_missing)[names(TSTdata_missing) == "congruency"] <- "trialtype"
TSTdata_missing$trialtype <- ifelse(TSTdata_missing$trialtype == "Congruent", "Repeat", "Switch")

A dataset containing data necessary to run examples of each function
}
\keyword{datasets}
