% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPsplithalf.R
\name{DPsplithalf}
\alias{DPsplithalf}
\title{Dot-Probe Split Half}
\usage{
DPsplithalf(data, RTmintrim, RTmaxtrim, incErrors = FALSE, conditionlist,
  halftype, no.iterations = 1, var.RT = "latency",
  var.condition = "blockcode", var.participant = "subject",
  var.correct = "correct", removelist = "")
}
\arguments{
\item{data}{specifies the raw dataset to be processed}

\item{RTmintrim}{specifies the lower cut-off point for RTs}

\item{RTmaxtrim}{specifies the maximum cut-off point for RTs}

\item{incErrors}{include incorrect trials?, defaults to FALSE}

\item{conditionlist}{sets conditions/blocks to be processed}

\item{halftype}{specifies the split method; "oddeven", "halfs", or "random"}

\item{no.iterations}{specifies the number of random splits to run}

\item{var.RT}{specifies the RT variable name in data}

\item{var.condition}{specifies the condition variable name in data}

\item{var.participant}{specifies the subject variable name in data}

\item{var.correct}{specifies the accuracy variable name in data}

\item{removelist}{specifies a list of participants to be removed}
}
\value{
Returns a data frame containing split-half reliability estimates for each condition specified.

splithalf returns the raw estimate

spearmanbrown returns the spearman-brown corrected estimate
}
\description{
This function calculates split half reliability estimates for Dot Probe data
}
