% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speedtestdata}
\alias{speedtestdata}
\title{Simulated data for runtime of splithalf package}
\format{
A data frame with 225 rows and 6 variables
}
\usage{
data(speedtestdata)
}
\description{
This simulation was run to estimate the relative runtimes for different possible combinations of sample sizes and trial numbers etc.
}
\details{
\itemize{
  \item{Sim}{codes for the simulation number}
  \item{sample_size}{codes for the sample size}
  \item{Number_of_conditions}{codes for the number of conditions run}
  \item{trials}{codes for the number of trials}
  \item{permutations}{codes for the number of permutations}
  \item{runtime}{codes for the runtime in seconds}
}
}
