% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splithalf_multiverse.R
\name{splithalf.multiverse}
\alias{splithalf.multiverse}
\title{Multiverse of data processing decisions on internal consistency reliability estimates.}
\usage{
splithalf.multiverse(input, specifications)
}
\arguments{
\item{input}{splithalf object or list of splithalf objects}

\item{specifications}{list of data processing specifications}
}
\value{
Returns a multiverse object containing the reliability estimates and dataframes from all data processing specifications provided
}
\description{
The (unofficial) function version name is "This function will let you get honey from a hornets nest"
}
\examples{
## see online documentation for examples
}
