% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testretest_multiverse.R
\name{testretest.multiverse}
\alias{testretest.multiverse}
\title{Multiverse of data processing decisions on test retest reliability estimates.}
\usage{
testretest.multiverse(
  data,
  specifications,
  test = "ICC2",
  outcome = "RT",
  score = "difference",
  var.participant = "subject",
  var.ACC = "correct",
  var.RT = "RT",
  var.time = "time",
  var.compare = "congruency",
  compare1 = "Congruent",
  compare2 = "Incongruent"
)
}
\arguments{
\item{data}{dataset}

\item{specifications}{list of data processing specifications}

\item{test}{correlation, ICC2, r ICC3}

\item{var.participant}{= "subject",}

\item{var.ACC}{= "correct",}

\item{var.RT}{set to internal consistency or test-retest}
}
\value{
Returns a multiverse object containing the reliability estimates and dataframes from all data processing specifications provided
}
\description{
The (unofficial) function version name is "This function will help you pay the troll toll"
}
\examples{
\dontrun{
## see online documentation for examples
https://github.com/sdparsons/splithalf
## also see https://psyarxiv.com/y6tcz

}
}
