% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testretest_multiverse.R
\name{testretest.multiverse}
\alias{testretest.multiverse}
\title{Multiverse of data processing decisions on test retest reliability estimates.}
\usage{
testretest.multiverse(
  input,
  specifications,
  test = "ICC2",
  var.participant = "subject",
  var.ACC = "correct",
  var.RT = "RT"
)
}
\arguments{
\item{input}{list of two datasets}

\item{specifications}{list of data processing specifications}

\item{test}{correlation, ICC2, r ICC3}

\item{var.participant}{= "subject",}

\item{var.ACC}{= "correct",}

\item{var.RT}{set to internal consistency or test-retest}
}
\value{
Returns a multiverse object containing the reliability estimates and dataframes from all data processing specifications provided
}
\description{
The (unofficial) function version name is "This function will help you pay the troll toll"
}
\examples{
## see online documentation for examples
}
